% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accuracy_per_class}
\alias{accuracy_per_class}
\title{TITLE accuracy_per_class}
\format{
A data frame with 17x3 
\describe{
  \item{class}{...}
  \item{counts}{..}
  ...
}
}
\usage{
accuracy_per_class
}
\description{
Description accuracy_per_class
}
\keyword{datasets}
