% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_for_tfidf}
\alias{data_for_tfidf}
\title{TITLE data_for_tfidf}
\format{
A data frame...
\describe{
  \item{class}{...}
  \item{counts}{..}
  ...
}
}
\usage{
data_for_tfidf
}
\description{
Description data_for_tfidf
}
\keyword{datasets}
