% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_best_estimators}
\alias{plot_best_estimators}
\title{Plot the best `Scikit-learn` (Python) estimators}
\usage{
plot_best_estimators(best_estimators)
}
\arguments{
\item{best_estimators}{The result of \code{\link{prep_best_estimators}}.}
}
\value{
A `ggplot2::geom_col`.
}
\description{
For internal use only!
}
