% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prep_all_pipeline_tuning_results}
\alias{prep_all_pipeline_tuning_results}
\title{Prepare table with all (hyper)parameter tunings}
\usage{
prep_all_pipeline_tuning_results(x)
}
\arguments{
\item{x}{A "tuning_results" data frame.}
}
\value{
A data frame.
}
\description{
For internal use only! Tidy up a "tuning_results" table from
[`pxtextmining`](https://github.com/CDU-data-science-team/pxtextmining).
}
