% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prep_best_estimators}
\alias{prep_best_estimators}
\title{Prepare table with the best (hyper)parameter tunings for each learner}
\usage{
prep_best_estimators(x)
}
\arguments{
\item{x}{A "tuning_results" data frame. See
\code{\link{prep_all_pipeline_tuning_results}}.}
}
\value{
A data frame.
}
\description{
For internal use only! Get the best-scoring (hyper)parameters for each
learner based on Class Balance Accuracy.
}
\note{
See \code{\link{prep_all_pipeline_tuning_results}}.
}
