% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{TITLE test_data}
\format{
A data frame...
\describe{
  \item{class}{...}
  \item{counts}{..}
  ...
}
}
\usage{
test_data
}
\description{
Description test_data
}
\keyword{datasets}
